requirements.txt
opencv-python
pycryptodome
tqdm
